refactor: shared StdioLoop for the JVM and Scala Native backends (TJC-2198)
`stdioLoop` and `writeLine` were byte-identical in JvmTransportBackend and
NativeTransportBackend (diff-clean, only scaladoc differed), and each backend
also re-derived the same session + stdout-semaphore wiring. All of it moves to
shared/.../transport/StdioLoop.scala:

  - `run`  — the dispatch loop: outbound drainer fiber, per-frame forked
             dispatch, EOF/interrupt teardown via `.ensuring`. Still callable
             over in-memory streams, which is what the lifecycle tests drive.
  - `serve` — the full lifecycle: mint the session, serialize stdout behind one
             writer, run the loop.

Each backend now carries only what genuinely differs on its platform: how stdin
becomes a line stream (`fromInputStream` + utf8Decode on the JVM;
`fromReader` + re-chunking on Native, which has no `fromInputStream`) and where
`randomId` gets entropy. JvmTransportBackend drops 93 → 45 lines, and
NativeTransportBackend 116 → 71.

MessageLoop keeps owning wire↔dispatch; StdioLoop owns the lifecycle around it,
which is exactly the layer that was duplicated.

JsTransportBackend is deliberately untouched: it drives Node's callback IO with
hand-rolled line framing and a UIO writeLine, so adopting this would be an I/O
rewrite rather than an extraction. Filed as a follow-up.

The Native lifecycle tests stay duplicated on purpose — the loop is shared now,
but its fiber semantics rest on Scala Native's threading and GC, so they re-run
the same contracts as a platform canary.

Verified: clean rebuild, 502/502 tasks green across all three platforms.
